requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
psycopg[binary]==3.2.6
python-dotenv==1.0.1
python-multipart==0.0.20
